\title{
Internal function for function \code{var.expDeath}}\name{function1}\alias{function1}\keyword{internal}\description{
This function was only meant to be used by the function \code{var.expDeath}. It provides many objects that are needed for the estimation of the variance of the expected number of deaths.
}
\usage{function1(x)}
\arguments{
  \item{x}{
\code{x} is an object of class \code{coxph}.
}
}\value{
The function \code{function1} returns a list containing many objects that are used by the function \code{var.expDeath}.}\author{
Denis Talbot, Thierry Duchesne, Jacques Brisson, Nathalie Vandal.
}\seealso{
\code{\link{var.expDeath}}
}
