\name{InferenceSMR-package}
\alias{InferenceSMR-package}
\alias{InferenceSMR}
\docType{package}
\title{
Inference about the standardized mortality ratio (SMR) when evaltuating the effect of a screening program on survival.}
\description{
The InferenceSMR package provides functions to make inference about the standardized mortality ratio (SMR) when evaltuating the effect of a screening program on survival. The package is based on methods described in Sasieni (2003) and Talbot et al. (2011). }
\details{
\tabular{ll}{
Package: \tab InferenceSMR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-05-22\cr
License: \tab GPL (>=) 2\cr
}

}
\author{
Denis Talbot, Thierry Duchesne, Jacques Brisson, Nathalie Vandal

Maintainer: Denis Talbot <denis.talbot@mat.ulaval.ca>
}
\references{
Sasieni P. (2003) \emph{On the expected number of cancer deaths during follow-up of an initially cancer-free cohort}.
Epidemiology, 14, 108-110.

Talbot, D., Duchesne, T., Brisson, J., Vandal, N. (2011) \emph{Variance estimation and confidence intervals for the standardized mortality ratio with application to the assessment of a cancer screening program}, Statistics in Medicine, 30, 3024-3037. 
}


\seealso{
\code{\link{est.expDeath}}, \code{\link{var.expDeath}}, \code{\link{inference.SMR}}, \code{\link{incidences}}, \code{\link{contrib}}, \code{\link{screening}}}

